add shared.collections StatesSet
