+++ regexy/shared/collections/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from .states_set import StatesSet
+
+__all__ = ['StatesSet']
